Create US_States_game_(gui).py
there's an image used in this code. download it from this link: https://i0.wp.com/whatsanswer.com/wp-content/uploads/2018/03/Blank-outline-map-of-the-United-States-611.jpg?fit=2000%2C1354&ssl=1

its been mentioned in the comments where you need to paste it. please name that image same as what is written in the code. and store this image in the same project folder where you have main.py file

store the image with filename as: blank_states_img.gif